requirements/common.txt
Django==5.0.2
django-graphql-jwt==0.4.0
graphene-django==2.16.0
psycopg2-binary==2.9
pytest==8.0.1
django-debug-toolbar==4.3.0
Pillow==10.2.0